# api/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Delivery
  description: Отвечает за учет курьеров, диспетчеризацию доставок, доставку
paths:
  /api/v1/orders:
    post:
      summary: Создать заказ
      description: Позволяет создать заказ с целью тестирования
      operationId: CreateOrder
      responses:
        '201':
          description: Успешный ответ
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Код ошибки
        message:
          type: string
          description: Текст ошибки